use faraday instead of http for requests
for some set of large responses, the http gem was raising an error:

> IOError: not opened for reading

i am not completely sure what the conditions for that failure are,
but request body size seems to be correlated with this occurring
consistently.

switching to faraday fixed the problem.

another thing worth noting is that this error was not being
rescued anywhere, which made it quite difficult to spot.
that is something that should be addressed regardless of this
patch.