`writekey` can be supplied via environment variable
A common strategy for managing sensitive data such as API keys is to set
them in the environment prior to starting a program that needs access to
them. Then the program can read from the environment variable at runtime.

This commit enables a similar strategy to work with libhoney out of the
box. If no `writekey` is supplied to `Libhoney::Client` it will check to
see if the `HONEYCOMB_WRITEKEY` environment variable is set. If it is it
will use that, otherwise it will use the empty string like it did
before.