If writekey/dataset not configured, warn and fall back to no-op...
Instead of waiting until event send time and then throwing an exception
for every event, which is at best noisy and unhelpful.

Strictly this is backward incompatible, since it was technically
possible for someone to leave writekey and dataset unspecified, and
instead set them via a builder or on individual events. (This doesn't
remove the latter capability, so it's still possible to send events to
multiple teams or datasets using the same `Libhoney::Client` - you just
have to pick a default writekey and dataset to initialize the Client.)