Time out HTTP requests (#34)
Consistent with libhoney-py and libhoney-go, this adds a 10 second timeout to each HTTP POST.

The http.rb library we're using supports granular timeouts for connect, write and read. For now, this just sets them all to 10 seconds. That means in principle a request could still last for 30 seconds without timing out (10s to connect, 10s to write the headers and event body, 10s to read the response headers). In practice, slowness in each phase is likely to have a different cause (e.g. connect slowness is probably the network, write/read probably shepherd hanging), so in most cases I'd expect this to time out after 10 seconds as desired.