Remove ArgumentError exceptions.
These lines are more harmful than useful. In the rare case that an event
is passed to the Transmission#send or Event#send_presampled methods
without certain fields. It is preferable to let the HTTP call fail than
to rudely throw an exception, which could happen in a critical path of a
request, causing a user visible error.