add test for closing transmission with a nil batch_thread
If the libhoney client is instantiated, but never started or an event
is not added, on process exit transmission throws an error when closing.

    NoMethodError (undefined method `join' for nil:NilClass) in 'close'

This behavior was seen in a Rails application running rake tasks that
initialize the entire app, but does not start or queue any events before the
process exits.

Co-authored-by: Dave Sullivan <dave@davesullivan.ca>
Signed-off-by: Robb Kidd <robbkidd@honeycomb.io>