deprecate array-form of proxy_config
The client gets instantiated once, so parse and display any warnings to the
user there where the API is public.

The parsing of the Array used by the http.rb library now assumes HTTP
connections to the proxy host for TCP tunneling of either HTTP or HTTPS.
If this is insufficient, http_proxy/https_proxy environment variables
should be used to configure the client instead, which prepares the
application for the deprecation removal also.

Signed-off-by: Robb Kidd <robbkidd@honeycomb.io>