require only the Status class from http.rb
Libhoney's Response object has historically been using a Status object
from the HTTP.rb library for its status_code property. In the context of
replacing the HTTP.rb library with Excon, this change keeps the gem
dependency on "http" so that we can make a very targeted "require" of
only the HTTP.rb's Status class. This targeted require avoids bringing in
most of the other HTTP.rb classes and their own dependencies,
particularly avoiding "DomainName" which is causing name conflicts
within the codebase of at least one of libhoney users.

It is probable that the Libhoney::Response object's status_code
interface will be deprecated from this fancy class with status
convenience methods to instead return a simple integer representing the
actual HTTP response status code for the client user to do with what
they will.