replace HTTP client library to reduce external dependencies (#81)
* Prefer excon over httprb

The `http` gem has a relatively large dependency graph, including the
`domain_name` gem which introduces a `DomainName` class into the global
namespace. Unfortunately, this can cause naming clashes which prevent
using Honeycomb.

Excon is actively maintained, supports a similar feature set, is
comparably performant, and has no dependencies.

* keep the http gem around for now to use its Response::Status class

Libhoney's Response object has historically been using a Status object
from the HTTP.rb library for its status_code property. We'll keep the gem
dependency on "http" so that we can make a very targeted "require" of
only the HTTP.rb's Status class. This targeted require avoids bringing in
most of the other HTTP.rb classes and their own dependencies,
particularly avoiding "DomainName" which is causing name conflicts
within the codebase of at least one of libhoney users.

It is probable that the Libhoney::Response object's status_code
interface will be deprecated in the near future from returning a fancy
Status class with convenience methods to instead return a simple
integer representing the HTTP response status code for the client user
to do with what they will.

* deprecate array-form of proxy_config

The client gets instantiated once, so parse and display any warnings to the
user during client creation where the API is public instead of down deep in
the private API of the Transmission HTTP client factory that gets called
repeatedly.

The parsing of the Array used by the http.rb library now assumes HTTP
connections to the proxy host for TCP tunneling of either HTTP or HTTPS.
If this is insufficient, http_proxy/https_proxy environment variables
should be used to configure the client instead, which prepares the
application for the deprecation removal also.

* add rudimentary debug logging to exception swallower

The handling of exceptions in this spot needs to have the option of
reporting to a human that an exception occurred to be able to debug
connection issues during event transmission. This is a super rudimentary
implementation to avoid the scope creep of adding an entire logging
interface.

Co-authored-by: Bernerd Schaefer <bj.schaefer@gmail.com>