make the success/error branches look more alike
Instantiating the Response within the call to enqueue_response() is a
bit easier for a human to read than tap()'ing a new Response instance to
enqueue from within the tap() block. This makes the two branches closer
to symmetrical.

Added a note to the status applied in the error condition. That's a big
part of what is different about error handling vs success handling.