Process single-error responses from the Batch API (#89)
An unsuccessful HTTP response from the Batch API endpoint has only
a single response for the entire event batch. That single response
needs to be multiplied by the number of batched events and enqueued
as the response for each. In the error case, only record an error response
for an event that successfully serialized and was included in the batch sent.

Instantiating the Response within the call to enqueue_response() is a
bit easier for a human to read than tap()'ing a new Response instance to
enqueue from within the tap() block. The success response branch was
updated to avoid using .tap().