update Client transmission param to accept a Class
This ought to provide a friendlier interface for declaring an
alternative transmission type that will still be configured with the
usual defaults.

The initialization of the transmission instance moves into Client
initialization via a call to a new setup_transmission(). This removes
the need for a lock and synchronizing with it on the send of every
event. A new instance of a Client will have an initialized Transmission.
If the transmission needs to change, you need a new Client.

setup_transmission() is necessarily complicated to handle the various
possibilities of what could be passed in. Tests have been added.

The Mock and Null transmissions have had their initialize methods
updated to accept (and ignore) any number of parameters. This lets
them be declared in the Client initialization like any other
transmission-like class.