Handle Timeout::Error in TransmissionClient
Release 1.19.0 added logging to the exception handler in
`batch_loop`. The problem with this is that under normal circumstances,
you will frequently find that the `batch_queue` is empty, and so
`Timeout.timeout(@send_frequency) { @batch_queue.pop }` will timeout and
raise a `Timeout::Error`. The error is logged every few hundred ms and
makes it practically impossible to work with `honeycomb-beeline` in the
console if you have LOG_LEVEL set to "debug" or "trace".