drop Ruby 2.3 support (#106)
Target Ruby 2.4 minimum, drop 2.3. Update Rubocop to last version that
supports Ruby 2.4. With the departure of Ruby 2.3, the CircleCI config
can use the newer Ruby images[1] intended to supercede the legacy
"circleci/ruby" images.

[1] https://circleci.com/developer/images/image/cimg/ruby

New Rubocop, New Appeasements:

* apply some config to Rubocop, turn on new rules, and regen TODO

* Lint/EmptyConditionalBody

    Avoid if branches without a body.

OK. Let's use the try operator (&.), then.

* Style/AccessorGrouping

    Group together all attr_accessor attributes.

OK

* Style/OptionalBooleanParameter

Disabling this rule for the close() method. Yes, a boolean parameter is
better represented by a keyword argument, but we don't want to meddle
with the backwards compatibility of this method right now.

* Style/GlobalStdStream

    Use $stderr instead of STDERR.

The first is a global variable. The latter is a constant. Changing the
global standard error destination is something we can assume ought to
be reflected in where the LogClient sends output.

So, OK.

* appease Rubocop: Style/NilLambda

    Use an empty lambda instead of always returning nil.

Empty lambdas return nil, so OK.

* Lint/EmptyClass

    Empty class detected.

CustomObject class sure was empty. Made a few changes:
- Added a .to_s method to the custom object test class. That's what the
  clean_data method uses when it is handed an object to turn into event
  data.
- Added another test case for the clean_data with object code path
  involving an object that raises an error if string coersion is
  attempted.
- Moved these test classes into the test case methods as instances of
  Class to make more clear that these are test objects.